Fix a race condition that causes the second batch to hang on start
Setting the max_prefetch=5 rather than 30 fixes this. I believe this has something to do with the creation of the BackgroundQueue for masks in the batch_size of the last batch of the validation set. The background execution of creating those masks likely overlaps with the start of the second batch. Maybe there's a problem with the data being changed out from under the process?