Replace the BackgroundGenerator with torch distribution generation on the GPU
This has a very positive effect on the performance. I no longer see the start and stop GPU usage. It is consistently at 93-97% usage. I verified that the probs and values are created directly on the GPU. My real dataset run is estimated to take 2hr40min on the first epoch, better than the previous best of 3hr30min.